Co-locate FrameColor var color extension, add textured bool
